Added raring by volume.

+++ Scraper.py
@@ -3,25 +3,19 @@ __author__ = 'Michael'
 
 import urllib.request
 import os
-from pathlib import Path
 from multiprocessing.pool import Pool
 
 
 def main():
-    home = os.getcwd()
     for i in range(1, 6):
-        download_dir = Path(os.getcwd() + "/Volume " + str(i))
-
-        if not download_dir.exists():
-            download_dir.mkdir()
-        os.chdir(download_dir.name)
 
         cover_url = "http://carciphona.com/_pages//" + str(i) + "/000.jpg"
         download_file(cover_url)
 
         urls = []
         for j in range(1, 200):
-            urls.append("http://carciphona.com/_pages//" + str(i) + "/" + str(j).zfill(3) + ".jpg")
+            url = "http://carciphona.com/_pages//" + str(i) + "/" + str(j).zfill(3) + ".jpg"
+            urls.append(url)
 
         # 2x as many processes as I have processors for better performance on more powerful systems, and because I may
         # get more speedup by utilising the time spent waiting for the website to respond.
@@ -29,19 +23,23 @@ def main():
         with Pool(12) as p:
             p.map(download_file, urls)
 
-        os.chdir(home)
+        # Move downloaded files into a .cbr archive.
+        rar_name = ("\"Volume " + str(i) + ".cbr\"")
+        os.system('rar m -m0 ' + rar_name + " *.jpg")
 
 
 def download_file(url):
     filename = url.split("/")[-1]
-    volume = url.split("/")[-2]
-    urllib.request.urlretrieve(url, filename)
 
-    if os.path.getsize(filename) < 25 * 1024:
-        os.remove(filename)
-        print("Page " + filename + " of volume " + volume + " does not exist, temp file removed.")
-    else:
-        print("Page " + filename + " of volume " + volume + " retrieved!")
+    try:
+        urllib.request.urlretrieve(url, filename)
+    except:
+        pass
+
+    if os.path.exists(filename):
+        if os.path.getsize(filename) < 25 * 1024:
+            os.remove(filename)
+
 
 if __name__ == "__main__":
     main()
